Extract Sheet helper functions into focused modules
Move pure computation functions from Sheet class into filter.ts,
clipboard.ts, and style-mutation.ts following the existing module
pattern (merging.ts, shifting.ts). Sheet class keeps state ownership
and store access, delegates computation to extracted functions.
No logic changes.
